removed unsused code

+++ store/models.py
@@ -14,7 +14,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=20, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=False)
 
     def __str__(self):
